feat(dashc): lower joined Figma variable bindings into the document
The compile request (ABI v2) gains the importer's joined binding rows:
one BoundVariable per boundVariables site, carrying the Figma node id,
the property path, the mode-qualified signal name, and the
mode-resolved value. The lowering maps the supported sites onto the
document binding tables — itemSpacing becomes a Gap binding, a solid
fill's color becomes four Fill channel bindings — and names every
unsupported site (figma.bindings.unsupported-property, warning: the
resolved literal still ships). A row targeting an unlowered node and an
inconsistent signal declaration are also named (P4).

A Custom binding transform reaching compile is a named error
(binding.custom-transform): the closure does not serialize, so the
document is refused, never emitted approximately.

The acceptance tests drive the whole path: the derived variables-bound
capture compiles to a .dsb whose tables load into the arena, attach_live
drives the dark card's mode-qualified signals into the committed scene,
and a dashlang-authored scene emits row-identical binding tables.

+++ crates/dashscene-core/src/load.rs
@@ -238,7 +238,7 @@ pub fn load_document(doc: &Document<'_>, arena: &mut Arena) -> u64 {
 /// value is `binding.unknown-channel` at the load gate, so it never
 /// reaches here (the same posture as the layout enums below).
 fn channel_of(channel: dashbuf::BindingChannel) -> Channel {
-    Channel::from_code(channel.0 as u8).unwrap_or_else(|| {
+    Channel::from_code(channel.0).unwrap_or_else(|| {
         unreachable!("unknown BindingChannel {channel:?}: rejected by the load gate (P4)")
     })
 }
